add range tests
