+++ packages/core/src/types/ranges.ts
@@ -1,0 +1,12 @@
+export type RangeFromOp = '>' | '>=';
+export type RangeToOp = '<' | '<=';
+
+export interface FromToRangeValues<T> {
+  readonly from?: T;
+  readonly to?: T;
+}
+
+export interface FromToRange<T> extends FromToRangeValues<T> {
+  readonly fromOp?: RangeFromOp;
+  readonly toOp?: RangeToOp;
+}
